Remove assign_address
